requirements.txt
flask
flask-socketio
flask-login
paho-mqtt
bcrypt
eventlet
pandas
numpy